+++ src/test-name-handler/TestNameHandler.ts
@@ -1,0 +1,9 @@
+export default interface TestNameHandler {
+  getCurrent: () => string | null
+
+  getPrevious: () => string | null
+
+  isNewTest: () => boolean
+
+  refresh: () => void
+}