+++ src/utils/listeners.ts
@@ -2,7 +2,7 @@ import { HTMLInputWithValue } from '../types'
 
 const TAG_NAME_DISALLOWED_BY_KEY_LISTENERS = ['textarea']
 const INPUT_ALLOWED_BY_KEY_LISTENERS = ['radio', 'select', 'checkbox', 'button']
-const KEYS_ALLOWED_BY_KEY_LISTENERS = ['tab']
+const KEYS_ALLOWED_BY_KEY_LISTENERS = ['tab', 'enter', 'numpadenter']
 
 export function isKeyAllowedByKeyListeners(keyCode: string): boolean {
   return KEYS_ALLOWED_BY_KEY_LISTENERS.includes(keyCode.toLowerCase())
@@ -12,7 +12,7 @@ export function isTargetAllowedByKeyListeners(target: EventTarget | null): boole
   if (target === null) return true
 
   const elementTarget = target as HTMLElement
-  if (TAG_NAME_DISALLOWED_BY_KEY_LISTENERS.includes(elementTarget.tagName.toLocaleLowerCase())) return false
+  if (TAG_NAME_DISALLOWED_BY_KEY_LISTENERS.includes(elementTarget.tagName.toLowerCase())) return false
 
   if (elementTarget.tagName.toLowerCase() === 'input') {
     const inputElementTarget = target as HTMLInputWithValue