+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Start a new session when the collector is run in a Cypress test context and the current test name is different from
+  the previous one.
+
 ### Deprecated
 
 ### Removed
@@ -23,7 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Stop collecting consecutive keydown actions targeting the same element (see [#9](https://github.com/Smartesting/gravity-data-collector/issues/9))
+- Stop collecting consecutive keydown actions targeting the same element (
+  see [#9](https://github.com/Smartesting/gravity-data-collector/issues/9))
 
 ### Fixed
 
@@ -49,8 +53,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Do not set `Origin` header when no source is provided. (see [#8](https://github.com/Smartesting/gravity-data-collector/issues/8))
-- Do not access directly `global.document`, leave it to a function dedicated if `global` is not defined. (see [#7](https://github.com/Smartesting/gravity-data-collector/issues/7))
+- Do not set `Origin` header when no source is provided. (
+  see [#8](https://github.com/Smartesting/gravity-data-collector/issues/8))
+- Do not access directly `global.document`, leave it to a function dedicated if `global` is not defined. (
+  see [#7](https://github.com/Smartesting/gravity-data-collector/issues/7))
 
 ## [2.1.3](https://github.com/Smartesting/gravity-data-collector/compare/v2.1.2...v2.1.3)